+++ sre_assistant/__init__.py
@@ -1,234 +1,11 @@
 # sre_assistant/__init__.py
 """
-SRE Assistant - ADK Agent Package
-暴露主代理供 A2A 調用，符合 2025 I/O A2A 增強協議
-參考 ARCHITECTURE.md 第 6.1 節
+SRE Assistant - ADK 代理套件。
+
+此檔案暴露主要的代理工作流程及其工廠函數，
+使其可被應用程式的其他部分（例如測試或伺服器入口點）匯入。
 """
 
-import os
-import uuid
-from datetime import datetime
-from collections import deque
-from typing import Literal, Optional, Dict, Deque
+from .workflow import SREWorkflow, create_workflow
 
-from fastapi import FastAPI
-import uvicorn
-from pydantic import BaseModel
-
-# ADK 和 A2A 相關導入
-# 注意：這些是假設的導入路徑，實際 SDK 可能不同
-# from google.adk.a2a import AgentCard, AgentCapabilities, AgentSkill, AgentSchema, SchemaVersion
-# from a2a_sdk.server import A2AStarletteApplication, DefaultRequestHandler, InMemoryTaskStore
-# from google.adk.a2a import Part, TextPart, new_artifact, completed_task, streaming_update
-# from a2a_sdk.exceptions import ServerError, UnsupportedOperationError, ValueError
-
-# 專案內部導入
-from .workflow import SREWorkflow
-from .a2a.protocol import StreamingChunk
-from .config.config_manager import config_manager, SessionBackend
-from .session.firestore_task_store import FirestoreTaskStore
-
-
-# --- 模擬 ADK/A2A SDK 元件 (因為實際 SDK 不可用) ---
-# 說明：由於我們無法訪問真實的 google.adk 或 a2a_sdk，
-# 我們在這裡創建最小化的模擬類別，以使程式碼結構完整且可運行。
-# 這些模擬類別反映了 ARCHITECTURE.md 中描述的功能。
-
-class MockA2AContext:
-    def __init__(self, task_id, context_id, message, streaming=True):
-        self.task_id = task_id
-        self.context_id = context_id
-        self.message = message
-        self._supports_streaming = streaming
-
-    def get_user_input(self):
-        return self.message
-
-    def supports_streaming(self):
-        return self._supports_streaming
-
-class MockEventQueue:
-    async def enqueue_event(self, event):
-        print(f"Event Queued: {event}")
-
-class MockA2AStarletteApplication:
-    def __init__(self, agent_card, http_handler):
-        self.app = FastAPI()
-        self.agent_card = agent_card
-        self.http_handler = http_handler
-        self._setup_routes()
-
-    def _setup_routes(self):
-        @self.app.get("/.well-known/agent.json")
-        async def get_agent_card():
-            return self.agent_card
-
-        @self.app.post("/execute")
-        async def execute(request: dict):
-            # 模擬執行流程
-            context = MockA2AContext(
-                task_id=str(uuid.uuid4()),
-                context_id=str(uuid.uuid4()),
-                message=request.get("query")
-            )
-            event_queue = MockEventQueue()
-            await self.http_handler.agent_executor.execute(context, event_queue)
-            return {"status": "execution_started"}
-
-    def build(self):
-        return self.app
-
-class DefaultRequestHandler:
-    def __init__(self, agent_executor, task_store):
-        self.agent_executor = agent_executor
-        self.task_store = task_store
-
-class InMemoryTaskStore:
-    pass
-
-class AgentCard(BaseModel):
-    name: str
-    version: str
-    description: str
-    url: str
-    # ... 其他元數據欄位
-
-# --- 核心實現 ---
-
-class StreamingHandler:
-    """處理 A2A Streaming 的 backpressure 和 idempotency"""
-    def __init__(self, event_queue: MockEventQueue, context: MockA2AContext):
-        self.event_queue = event_queue
-        self.context = context
-        self.buffer: Deque[StreamingChunk] = deque(maxlen=100)  # 用於 backpressure 的緩衝區
-        self.seen_tokens: set[str] = set()  # 用於 idempotency
-
-    async def handle_with_flow_control(self, chunk: StreamingChunk):
-        """處理單個 chunk，包含流量控制和冪等性檢查"""
-        if chunk.idempotency_token in self.seen_tokens:
-            print(f"Skipping duplicate chunk: {chunk.idempotency_token}")
-            return  # 防止重複處理
-
-        self.seen_tokens.add(chunk.idempotency_token)
-        self.buffer.append(chunk)
-
-        if len(self.buffer) >= self.buffer.maxlen:
-            print("Buffer full, pausing producer...")
-            # 在真實世界中，這裡會發送一個信號給生產者暫停發送
-
-        # 模擬發送 streaming 事件
-        await self.event_queue.enqueue_event({
-            "event_type": "streaming_update",
-            "task_id": self.context.task_id,
-            "chunk": chunk.model_dump_json()
-        })
-
-class SREAssistantExecutor:
-    """SRE Assistant 執行器，用於處理 A2A 請求"""
-    def __init__(self, agent: SREWorkflow):
-        self.agent = agent
-
-    async def execute(self, context: MockA2AContext, event_queue: MockEventQueue):
-        query = context.get_user_input()
-        if context.supports_streaming():
-            await self._execute_with_streaming(context, event_queue, query)
-        else:
-            await self._execute_batch(context, event_queue, query)
-
-    async def _execute_with_streaming(self, context: MockA2AContext, event_queue: MockEventQueue, query: str):
-        """支援 streaming 的執行模式"""
-        handler = StreamingHandler(event_queue, context)
-        task_id = context.task_id
-
-        # self.agent.execute_streaming 是假設的方法，目前 SRECoordinator 中不存在
-        # 我們將模擬它的行為
-        print("Starting streaming execution...")
-        i = 0
-        async for chunk_data in self.agent.execute_streaming(query):
-            i += 1
-            chunk = StreamingChunk(
-                type="partial_result",
-                partial_result=chunk_data,
-                progress=(i / 5.0), # 假設總共有 5 個塊
-                idempotency_token=f"sre_stream_{task_id}_{i}"
-            )
-            await handler.handle_with_flow_control(chunk)
-
-        # 發送最終完成事件
-        final_chunk = StreamingChunk(
-            type="final_result",
-            final_result={"status": "completed"},
-            progress=1.0,
-            idempotency_token=f"sre_stream_{task_id}_final"
-        )
-        await handler.handle_with_flow_control(final_chunk)
-
-        await event_queue.enqueue_event({
-            "event_type": "completed_task",
-            "task_id": context.task_id
-        })
-
-    async def _execute_batch(self, context: MockA2AContext, event_queue: MockEventQueue, query: str):
-        """批次處理模式"""
-        # self.agent.execute 是假設的方法
-        result = await self.agent.execute(query)
-        await event_queue.enqueue_event({
-            "event_type": "completed_task",
-            "task_id": context.task_id,
-            "result": result
-        })
-
-def create_agent_card(host="0.0.0.0", port=8080) -> AgentCard:
-    """建立 AgentCard，定義 SRE Assistant 的能力"""
-    return AgentCard(
-        name="sre_assistant",
-        version="1.0.0",
-        description="AI-powered SRE Assistant for automated operations and maintenance",
-        url=os.getenv("HOST_OVERRIDE") or f"http://{host}:{port}/",
-        # ... 其他來自 ARCHITECTURE.md 的詳細資訊可以在此添加
-    )
-
-# --- 應用程式設置 ---
-
-def get_task_store():
-    """根據配置創建並返回對應的 TaskStore 實例。"""
-    if config_manager.config.session_backend == SessionBackend.FIRESTORE:
-        project_id = config_manager.config.firestore_project_id
-        if not project_id:
-            # 在 GCP 環境中，我們可以嘗試自動檢測 Project ID
-            project_id = os.getenv("GOOGLE_CLOUD_PROJECT")
-            if not project_id:
-                raise ValueError(
-                    "Firestore backend requires 'firestore_project_id' in config "
-                    "or GOOGLE_CLOUD_PROJECT environment variable."
-                )
-        return FirestoreTaskStore(
-            project_id=project_id,
-            collection=config_manager.config.firestore_collection
-        )
-    else: # 默認返回 InMemoryTaskStore
-        return InMemoryTaskStore()
-
-# 實例化主工作流程
-sre_agent = SREWorkflow(config=config_manager.config.model_dump())
-
-# 組合 A2A 服務
-executor = SREAssistantExecutor(agent=sre_agent)
-request_handler = DefaultRequestHandler(
-    agent_executor=executor,
-    task_store=get_task_store(),
-)
-server = MockA2AStarletteApplication(
-    agent_card=create_agent_card(),
-    http_handler=request_handler
-)
-app = server.build()
-
-# --- ADK 套件級別的暴露 ---
-# 這些變數可以被 ADK 框架發現
-__agent__ = sre_agent
-__version__ = "1.0.0"
-
-# --- 主函數 (用於直接運行) ---
-if __name__ == "__main__":
-    uvicorn.run(app, host="0.0.0.0", port=8080)
+__all__ = ["SREWorkflow", "create_workflow"]
