+++ sre_assistant/sub_agents/remediation/__init__.py
@@ -1,2 +1,6 @@
 # sre_assistant/sub_agents/remediation/__init__.py
-# 說明：此檔案將 remediation 目錄標記為一個 Python 子套件。
+# 該檔案將 remediation 目錄標記為一個 Python 子套件，
+# 並暴露其關鍵組件。
+
+from .agent import RemediationAgent
+from .dispatcher_agent import SREIntelligentDispatcher
